+++ airbyte-integrations/connectors/source-todoist-backup/acceptance-test-config.yml
@@ -1,6 +1,7 @@
-# See [Connector Acceptance Tests](https://docs.airbyte.com/connector-development/testing-connectors/connector-acceptance-tests-reference)
+# See [Source Acceptance Tests](https://docs.airbyte.com/connector-development/testing-connectors/source-acceptance-tests-reference)
 # for more information about how to configure these tests
 connector_image: airbyte/source-todoist:dev
+test_strictness_level: low
 acceptance_tests:
   spec:
     tests:
@@ -19,20 +20,19 @@ acceptance_tests:
       - config_path: "secrets/config.json"
         configured_catalog_path: "integration_tests/configured_catalog.json"
         empty_streams: []
-# TODO uncomment this block to specify that the tests should assert the connector outputs the records provided in the input file a file
-#        expect_records:
-#          path: "integration_tests/expected_records.jsonl"
-#          extra_fields: no
-#          exact_order: no
-#          extra_records: yes
-  incremental: 
+  # TODO uncomment this block to specify that the tests should assert the connector outputs the records provided in the input file a file
+  #        expect_records:
+  #          path: "integration_tests/expected_records.txt"
+  #          extra_fields: no
+  #          exact_order: no
+  #          extra_records: yes
+  incremental:
     bypass_reason: "This connector does not implement incremental sync"
-# TODO uncomment this block this block if your connector implements incremental sync: 
-#    tests:
-#      - config_path: "secrets/config.json"
-#        configured_catalog_path: "integration_tests/configured_catalog.json"
-#        future_state:
-#          future_state_path: "integration_tests/abnormal_state.json"
+  # TODO uncomment this block this block if your connector implements incremental sync:
+  #    tests:
+  #      - config_path: "secrets/config.json"
+  #        configured_catalog_path: "integration_tests/configured_catalog.json"
+  #        future_state_path: "integration_tests/abnormal_state.json"
   full_refresh:
     tests:
       - config_path: "secrets/config.json"
