+++ airbyte-integrations/connectors/source-todoist-backup/integration_tests/configured_catalog.json
@@ -1,0 +1,34 @@
+{
+  "streams": [
+    {
+      "cursor_field": null,
+      "destination_sync_mode": "append",
+      "primary_key": null,
+      "stream": {
+        "default_cursor_field": null,
+        "json_schema": {},
+        "name": "tasks",
+        "namespace": null,
+        "source_defined_cursor": null,
+        "source_defined_primary_key": [["id"]],
+        "supported_sync_modes": ["full_refresh"]
+      },
+      "sync_mode": "full_refresh"
+    },
+    {
+      "cursor_field": null,
+      "destination_sync_mode": "append",
+      "primary_key": null,
+      "stream": {
+        "default_cursor_field": null,
+        "json_schema": {},
+        "name": "projects",
+        "namespace": null,
+        "source_defined_cursor": null,
+        "source_defined_primary_key": [["id"]],
+        "supported_sync_modes": ["full_refresh"]
+      },
+      "sync_mode": "full_refresh"
+    }
+  ]
+}
