+++ airbyte-integrations/connectors/source-todoist-backup/integration_tests/invalid_config.json
@@ -1,0 +1,3 @@
+{
+  "token": "INVALID TOKEN"
+}
