+++ airbyte-integrations/connectors/source-todoist-backup/integration_tests/sample_config.json
@@ -1,0 +1,3 @@
+{
+  "token": "VALID TOKEN"
+}
