+++ airbyte-integrations/connectors/source-todoist-backup/integration_tests/sample_state.json
@@ -1,0 +1,5 @@
+{
+  "todo-stream-name": {
+    "todo-field-name": "value"
+  }
+}
