+++ airbyte-integrations/connectors/source-todoist-backup/metadata.yaml
@@ -1,0 +1,24 @@
+data:
+  connectorSubtype: api
+  connectorType: source
+  definitionId: 7d272065-c316-4c04-a433-cd4ee143f83e
+  dockerImageTag: 0.1.0
+  dockerRepository: airbyte/source-todoist
+  githubIssueLabel: source-todoist
+  icon: todoist.svg
+  license: MIT
+  name: Todoist
+  registries:
+    cloud:
+      enabled: true
+    oss:
+      enabled: true
+  releaseStage: alpha
+  documentationUrl: https://docs.airbyte.com/integrations/sources/todoist
+  tags:
+    - language:python
+  ab_internal:
+    sl: 100
+    ql: 100
+  supportLevel: community
+metadataSpecVersion: "1.0"
