+++ airbyte-integrations/connectors/source-todoist-backup/requirements.txt
@@ -1,0 +1,1 @@
+-e .
