+++ airbyte-integrations/connectors/source-todoist-backup/setup.py
@@ -6,14 +6,14 @@
 from setuptools import find_packages, setup
 
 MAIN_REQUIREMENTS = [
-    "airbyte-cdk~=0.1",
+    "airbyte-cdk~=0.2",
 ]
 
 TEST_REQUIREMENTS = [
     "requests-mock~=1.9.3",
-    "pytest~=6.2",
-    "pytest-mock~=3.6.1",
-    "connector-acceptance-test",
+    "pytest~=6.1",
+    "pytest-mock~=3.6",
+    "requests_mock~=1.8",
 ]
 
 setup(
