+++ airbyte-integrations/connectors/source-todoist-backup/source_todoist/schemas/projects.json
@@ -1,0 +1,43 @@
+{
+  "$schema": "http://json-schema.org/schema#",
+  "additionalProperties": true,
+  "type": "object",
+  "properties": {
+    "color": {
+      "type": ["null", "string"]
+    },
+    "comment_count": {
+      "type": ["null", "integer"]
+    },
+    "id": {
+      "type": ["null", "string"]
+    },
+    "is_favorite": {
+      "type": ["null", "boolean"]
+    },
+    "is_inbox_project": {
+      "type": ["null", "boolean"]
+    },
+    "is_shared": {
+      "type": ["null", "boolean"]
+    },
+    "is_team_inbox": {
+      "type": ["null", "boolean"]
+    },
+    "name": {
+      "type": ["null", "string"]
+    },
+    "order": {
+      "type": ["null", "integer"]
+    },
+    "parent_id": {
+      "type": ["null", "string"]
+    },
+    "url": {
+      "type": ["null", "string"]
+    },
+    "view_style": {
+      "type": ["null", "string"]
+    }
+  }
+}
