+++ airbyte-integrations/connectors/source-todoist-backup/source_todoist/schemas/tasks.json
@@ -1,0 +1,82 @@
+{
+  "$schema": "http://json-schema.org/schema#",
+  "additionalProperties": true,
+  "type": "object",
+  "properties": {
+    "assignee_id": {
+      "type": ["null", "string"]
+    },
+    "assigner_id": {
+      "type": ["null", "string"]
+    },
+    "comment_count": {
+      "type": ["null", "integer"]
+    },
+    "content": {
+      "type": ["null", "string"]
+    },
+    "created_at": {
+      "type": ["null", "string"]
+    },
+    "creator_id": {
+      "type": ["null", "string"]
+    },
+    "description": {
+      "type": ["null", "string"]
+    },
+    "due": {
+      "anyOf": [
+        {
+          "type": ["null", "object"]
+        },
+        {
+          "properties": {
+            "date": {
+              "type": ["null", "string"]
+            },
+            "is_recurring": {
+              "type": ["null", "boolean"]
+            },
+            "lang": {
+              "type": ["null", "string"]
+            },
+            "string": {
+              "type": ["null", "string"]
+            }
+          },
+          "type": ["null", "object"]
+        }
+      ]
+    },
+    "id": {
+      "type": ["null", "string"]
+    },
+    "is_completed": {
+      "type": ["null", "boolean"]
+    },
+    "labels": {
+      "type": ["null", "array"],
+      "items": {
+        "type": ["null", "string"]
+      }
+    },
+    "order": {
+      "type": ["null", "integer"]
+    },
+    "parent_id": {
+      "type": ["null", "string"]
+    },
+    "priority": {
+      "type": ["null", "integer"]
+    },
+    "project_id": {
+      "type": ["null", "string"]
+    },
+    "section_id": {
+      "type": ["null", "string"]
+    },
+    "url": {
+      "type": ["null", "string"]
+    }
+  }
+}
