+++ airbyte-integrations/connectors/source-todoist/integration_tests/configured_catalog.json
@@ -1,34 +1,22 @@
 {
   "streams": [
     {
-      "cursor_field": null,
-      "destination_sync_mode": "append",
-      "primary_key": null,
       "stream": {
-        "default_cursor_field": null,
-        "json_schema": {},
         "name": "tasks",
-        "namespace": null,
-        "source_defined_cursor": null,
-        "source_defined_primary_key": [["id"]],
+        "json_schema": {},
         "supported_sync_modes": ["full_refresh"]
       },
-      "sync_mode": "full_refresh"
+      "sync_mode": "full_refresh",
+      "destination_sync_mode": "overwrite"
     },
     {
-      "cursor_field": null,
-      "destination_sync_mode": "append",
-      "primary_key": null,
       "stream": {
-        "default_cursor_field": null,
-        "json_schema": {},
         "name": "projects",
-        "namespace": null,
-        "source_defined_cursor": null,
-        "source_defined_primary_key": [["id"]],
+        "json_schema": {},
         "supported_sync_modes": ["full_refresh"]
       },
-      "sync_mode": "full_refresh"
+      "sync_mode": "full_refresh",
+      "destination_sync_mode": "overwrite"
     }
   ]
 }
