+++ airbyte-integrations/connectors/source-todoist/integration_tests/sample_config.json
@@ -1,3 +1,3 @@
 {
-  "token": "VALID TOKEN"
+  "token": "API_KEY"
 }
