+++ airbyte-integrations/connectors/source-todoist/source_todoist/manifest.yaml
@@ -1,0 +1,62 @@
+version: "0.29.0"
+
+definitions:
+  selector:
+    type: RecordSelector
+    extractor:
+      type: DpathExtractor
+      field_path: []
+  requester:
+    type: HttpRequester
+    url_base: "https://api.todoist.com/rest/v2"
+    http_method: "GET"
+    authenticator:
+      type: BearerAuthenticator
+      api_token: "{{ config['token'] }}"
+  retriever:
+    type: SimpleRetriever
+    record_selector:
+      $ref: "#/definitions/selector"
+    paginator:
+      type: NoPagination
+    requester:
+      $ref: "#/definitions/requester"
+  base_stream:
+    type: DeclarativeStream
+    retriever:
+      $ref: "#/definitions/retriever"
+  tasks_stream:
+    $ref: "#/definitions/base_stream"
+    name: "tasks"
+    $parameters:
+      path: "/tasks"
+  projects_stream:
+    $ref: "#/definitions/base_stream"
+    name: "projects"
+    $parameters:
+      path: "/projects"
+
+
+streams:
+  - "#/definitions/tasks_stream"
+  - "#/definitions/projects_stream"
+
+check:
+  type: CheckStream
+  stream_names:
+    - "tasks"
+    - "projects"
+
+spec:
+  type: Spec
+  documentation_url: https://docs.airbyte.com/integrations/sources/source-todolist
+  connection_specification:
+    title: Source Todolist Spec
+    type: object
+    required:
+      - token
+    additionalProperties: true
+    properties:
+      token:
+        type: string
+        description: API authorization bearer token for authenticating the API
